fixed tests with const assertions
